docs/examples/google-core-iot/requirements.txt
google-api-python-client
pytz==2018.9
requests==2.20.1
cryptography==2.4.2